Add permissions and fix user getting

+++ src/api/v1/car/views.py
@@ -1,9 +1,11 @@
 from django.shortcuts import get_object_or_404
 from rest_framework import viewsets
 from rest_framework.decorators import action
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
 from rest_framework import status
 
+from src.api.auth.permissions import IsOwnerOrReadOnlyAndCreate
 from src.apps.car.models import Car
 from src.apps.car.serializers import CarSerializer
 from src.apps.comment.models import Comment
@@ -11,10 +13,13 @@ from src.apps.comment.serializers import CommentSerializer
 
 
 class CarViewSet(viewsets.ViewSet):
-    """VewSet для сущности "Автомобиль", методы: list, retrieve, create, destroy, update"""
+    """
+    VewSet для сущности "Автомобиль", методы: list, retrieve, create, destroy, update, create_comment, get_comments
+    """
 
     serializer_class = CarSerializer
     queryset = Car.objects.all()
+    permission_classes = (IsAuthenticatedOrReadOnly, IsOwnerOrReadOnlyAndCreate)
 
     def list(self, request):
         serializer = self.serializer_class(self.queryset, many=True)
@@ -26,6 +31,7 @@ class CarViewSet(viewsets.ViewSet):
         return Response(serializer.data)
 
     def create(self, request):
+        request.data["owner"] = request.user.id
         serializer = self.serializer_class(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -33,6 +39,10 @@ class CarViewSet(viewsets.ViewSet):
 
     def update(self, request, pk=None):
         instance = get_object_or_404(self.queryset, pk=pk)
+
+        self.check_object_permissions(request, instance)
+
+        request.data["owner"] = request.user.id
         serializer = self.serializer_class(data=request.data, instance=instance)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -40,18 +50,23 @@ class CarViewSet(viewsets.ViewSet):
 
     def destroy(self, request, pk):
         instance = get_object_or_404(self.queryset, pk=pk)
+
+        self.check_object_permissions(request, instance)
+
         instance.delete()
         return Response(None)
 
     @action(detail=True, methods=["get", "post"], url_path="comments")
     def comments(self, request, pk=None):
         if request.method == "GET":
-            comments = Comment.objects.filter(car_id=pk)
+            comments = get_object_or_404(self.queryset, pk=pk).comments.all()
             serializer = CommentSerializer(comments, many=True)
             return Response(serializer.data)
 
         if request.method == "POST":
+            get_object_or_404(self.queryset, pk=pk)
             request.data["car"] = pk
+            request.data["author"] = request.user.id
             serializer = CommentSerializer(data=request.data)
             serializer.is_valid(raise_exception=True)
             serializer.save()
